rename id minp
